gridify does not give a coherent grid
When using gridify to convert from hex to square grid, ebsd.dx and dy become the numbers indicated in "unit_cell" (for example 5), but the spacing of the grid, ebsd.x and ebsd.y is in different steps (here 4.996). This leads to problems when import the data (tested with Atex), which is why I compute a new grid in the export function now. Please check whether this is ok with you.