Update OptimalKernel.m
Some grammatical edits, smoothing out the sentence structure, and clarifying that the calculations mentioned are settings for the calcKernel function that can be set with flags.  I'm not sure about sampling size being defined in line 82 - should it match the number of orientations sampled (i.e. the length of ori)?  Added more comments to the code.  Am I right about the units for error on line 111?