Update ClusterDemo.m
Added intro to the demo at top, added comments throughout the code.  plotCluster(r,cId) command seems to be redundant- is it necessary?  If so, it should be listed at the top of the page (before it's called) with some instructions on how to save it (a lot of people just copy and paste the code in order starting from the top of the webpage).  Added semicolons to remove some verbose text.