Update calcVariantGraph.m
The patch is meant to fix behavior in the case there is information in job.votes of merged variant indices. Previously, the function limits the variants in these cases to the best fitting variant (of 12) and the second best fitting variant (of 12), typically leading to a contest of indices with a twinning relationship. Now, the function limits the variants in the graph to the best fitting variant and its closest variant neighbor (of 24).