Match map plot direction to a vector3d
Useful for when you want to plot or measure something wrt e.g. 'north' or 'west' on the map but don't want to hard code it as vector3d.X or Y, might be useful when looking at the same sample with different EBSD systems. 

I'm not sure which folder this function could go into