bug fix
the methods (S2Fun.grad and S2FunHarmonic.grad) should coincide. Also the evaluation of the sF.grad vector field in some points should coincide with the direct computation of the gradient at this points.
Moreover we should compute the tangent space right, since perp do not yield elements of the tangent space in case of lists.