bug fix
May get an vector by evaluating on quadratureSO3Grid also for SO3Fun's of length 1.
This is necessary if u want to do quadrature of an SO3FunHandle which consists of multiple combined SO3Fun's with different sizes, for instant in .*. In this case the size of the inner evaluation methods have to be the same.
One option could be to insert rot(:), but this would destroy the structure (class) of the quadratureSO3Grid. However we want to use evalEuispacedFFT and quadratureSO3Grid.eval to speed up the evaluation process.
Hence we added the option evalShape with the flag 'vector' such that the evaluation methods yields an output of size length(rot)x1.