wignerTrafo transforms harmonic to Fourier coefficents
Therefore it uses the wignerTrafomex routine, which infact is only an part of the wigner transform, since the missing coefficients have to be reconstructed afterwards.
Moreover we have to to multiply the Fourier coefficients by i^(k-l) since this is part of the Wigner transform. Therefore it is not anymore neccessary to shift the Euler angles at the beginning of the eval function.
The evalSectionsFFT method now is useable if the bandwidth is very high in comparison to the Grid size. Hence this could be a powerfull tool by plotting SO3FunHarmonics of very high bandwidths.