restructure the SO3FunRBF methods
There is no method which is called SO3FunRBF.quadrature.
We delete the SO3Fun.interpolate method (it is merged into the approximation method)
Now (similar as in SO3FunHarmonic) there is only the SO3FunRBF.approximation method where i can input a @SO3Fun or data points (nodes,values) to approximate a suitable unimodalODF.
If the input is of class @SO3Fun we can again also use the constructor SO3FunRBF for approximation.