Simplify the max routines
Therefore the SO3Fun/max method checks and computes the max in case of vector valued functions.
So we do not need to add this multivariate procedure in every single max function.
This requires access to the subSet routine also from the SO3Fun class. Hence it can not be private anymore.