basic computation methods for VectorFields.
The tangent vectors only have the structure of a vector space. That means, we have summation and scalar multiplication, but no multiplication of tangent vectors and no elementwise summation with numbers. This operations would be inconsistent with changing between the two different tangent space representations (left <-> right).
This transfers to the vector fields. Hence SO3Vectorfields can be added only with themselves. They can multiplied with scalar SO3Funs and scalar values.